+++ test/bigtest/tests/loan-policy/loan-policy-form-test.js
@@ -1,8 +1,7 @@
 import { beforeEach, describe, it } from '@bigtest/mocha';
 import { expect } from 'chai';
 
-// eslint-disable-next-line import/extensions, import/no-extraneous-dependencies, import/no-unresolved
-import translation from '@folio/circulation/translations/ui-circulation/en.json';
+import translation from '../../../../translations/ui-circulation/en';
 
 import {
   shortTermLoansOptions,
@@ -13,9 +12,8 @@ import setupApplication from '../../helpers/setup-application';
 import LoanPolicyForm from '../../interactors/loan-policy/loan-policy-form';
 import LoanPolicyDetail from '../../interactors/loan-policy/loan-policy-detail';
 import {
-  getRequiredLabel,
   getOptionsRepresentation,
-} from '../../helpers/messageСonverters';
+} from '../../helpers/messageConverters';
 
 describe('LoanPolicyForm', () => {
   setupApplication();
@@ -100,9 +98,7 @@ describe('LoanPolicyForm', () => {
         });
 
         it('should have proper label', () => {
-          expect(LoanPolicyForm.aboutSection.policyName.label).to.equal(
-            getRequiredLabel(translation['settings.loanPolicy.policyName'])
-          );
+          expect(LoanPolicyForm.aboutSection.policyName.label).to.contain(translation['settings.loanPolicy.policyName']);
         });
       });
 
@@ -170,7 +166,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.loansSection.loanProfile.label).to.equal(`${translation['settings.loanPolicy.loanProfile']} *`);
+              expect(LoanPolicyForm.loansSection.loanProfile.label).to.contain(translation['settings.loanPolicy.loanProfile']);
             });
           });
 
@@ -196,9 +192,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.loansSection.loanPeriod.label).to.equal(
-                getRequiredLabel(translation['settings.loanPolicy.loanPeriod'])
-              );
+              expect(LoanPolicyForm.loansSection.loanPeriod.label).to.contain(translation['settings.loanPolicy.loanPeriod']);
             });
           });
 
@@ -250,9 +244,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.loansSection.fixedDueDateScheduleId.label).to.equal(
-                getRequiredLabel(translation['settings.loanPolicy.fDDS'])
-              );
+              expect(LoanPolicyForm.loansSection.fixedDueDateScheduleId.label).to.contain(translation['settings.loanPolicy.fDDS']);
             });
           });
 
@@ -274,7 +266,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.loansSection.closedDueDateMgmt.label).to.equal(`${translation['settings.loanPolicy.closedDueDateMgmt']} *`);
+              expect(LoanPolicyForm.loansSection.closedDueDateMgmt.label).to.contain(translation['settings.loanPolicy.closedDueDateMgmt']);
             });
           });
 
@@ -321,9 +313,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.loansSection.openingTimeOffset.label).to.equal(
-                getRequiredLabel(translation['settings.loanPolicy.openingTimeOffset'])
-              );
+              expect(LoanPolicyForm.loansSection.openingTimeOffset.label).to.contain(translation['settings.loanPolicy.openingTimeOffset']);
             });
           });
         });
@@ -459,9 +449,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.renewalsSection.numRenewalsAllowed.label).to.equal(
-                getRequiredLabel(translation['settings.loanPolicy.numRenewalsAllowed'])
-              );
+              expect(LoanPolicyForm.renewalsSection.numRenewalsAllowed.label).to.contain(translation['settings.loanPolicy.numRenewalsAllowed']);
             });
           });
 
@@ -507,9 +495,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.renewalsSection.renewFrom.label).to.equal(
-                `${translation['settings.loanPolicy.renewFrom']} *`
-              );
+              expect(LoanPolicyForm.renewalsSection.renewFrom.label).to.contain(translation['settings.loanPolicy.renewFrom']);
             });
           });
 
@@ -618,9 +604,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.renewalsSection.alternateLoanPeriodRenewals.label).to.equal(
-                getRequiredLabel(translation['settings.loanPolicy.alternateLoanPeriodRenewals'])
-              );
+              expect(LoanPolicyForm.renewalsSection.alternateLoanPeriodRenewals.label).to.contain(translation['settings.loanPolicy.alternateLoanPeriodRenewals']);
             });
           });
         });
@@ -663,7 +647,7 @@ describe('LoanPolicyForm', () => {
             });
 
             it('should have proper label', () => {
-              expect(LoanPolicyForm.renewalsSection.alternateFixedDueDateSchedule.label).to.equal(
+              expect(LoanPolicyForm.renewalsSection.alternateFixedDueDateSchedule.label).to.contain(
                 translation['settings.loanPolicy.altFDDSforRenewals']
               );
             });