+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.10.0 (IN PROGRESS)
 
 * Retrieve 1000 elements on Settings pages instead of 10. Refs UICIRC-302.
+* Add missing tests for request policy. Part of UICIRC-293.
 
 ## [1.9.0](https://github.com/folio-org/ui-circulation/tree/v1.9.0) (2019-07-25)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.8.0...v1.9.0)